third_party/blink/web_tests/http/tests/eventsource/workers/eventsource-url-attribute-expected.txt

[Worker] Only .url should work, previously supported .URL should not.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: ../script-tests/eventsource-url-attribute.js
PASS [Worker] source.url is "http://127.0.0.1:8000/eventsource/resources/event-stream.php"
PASS [Worker] source.URL is undefined.
PASS successfullyParsed is true

TEST COMPLETE
